requirements.txt
# Core dependencies
fastapi==0.115.12
uvicorn[standard]==0.32.1
pydantic==2.10.4
pydantic-settings==2.6.1

# AI and ML
openai==1.83.0
langchain==0.3.25
langchain-openai==0.3.19
langchain-community==0.3.24
langchain-core==0.3.63
langchain-text-splitters==0.3.8

# Vector database
faiss-cpu==1.11.0
numpy==2.0.2
pandas==2.2.3
scikit-learn==1.6.2

# Document processing
python-docx==1.1.2
pypdf==4.2.0
fpdf==1.7.2
python-magic==0.4.27

# Web and API
python-multipart==0.0.20
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.28.1

# Configuration and utilities
python-dotenv==1.0.1
PyYAML==6.0.2
jinja2==3.1.6

# Development and testing
pytest==8.2.2
pytest-asyncio==0.24.0
black==24.12.0
flake8==7.1.1
mypy==1.13.0

# UI (optional)
streamlit==1.36.0 